swagger 对 long 类型识别有问题，修改为 integer 了

+++ grpc2http-gateway-core/src/main/java/com/qianxunclub/grpc2httpgateway/utils/FieldTypeEnum.java
@@ -8,10 +8,10 @@ public enum FieldTypeEnum {
 
     DOUBLE(Descriptors.FieldDescriptor.JavaType.DOUBLE, Double.class.getSimpleName()),
     FLOAT(Descriptors.FieldDescriptor.JavaType.FLOAT, Float.class.getSimpleName()),
-    INT64(Descriptors.FieldDescriptor.JavaType.LONG, Long.class.getSimpleName()),
-    UINT64(Descriptors.FieldDescriptor.JavaType.LONG, Long.class.getSimpleName()),
+    INT64(Descriptors.FieldDescriptor.JavaType.LONG, Integer.class.getSimpleName()),
+    UINT64(Descriptors.FieldDescriptor.JavaType.LONG, Integer.class.getSimpleName()),
     INT32(Descriptors.FieldDescriptor.JavaType.INT, Integer.class.getSimpleName()),
-    FIXED64(Descriptors.FieldDescriptor.JavaType.LONG, Long.class.getSimpleName()),
+    FIXED64(Descriptors.FieldDescriptor.JavaType.LONG, Integer.class.getSimpleName()),
     FIXED32(Descriptors.FieldDescriptor.JavaType.INT, Integer.class.getSimpleName()),
     BOOL(Descriptors.FieldDescriptor.JavaType.BOOLEAN, Boolean.class.getSimpleName()),
     STRING(Descriptors.FieldDescriptor.JavaType.STRING, String.class.getSimpleName()),
@@ -21,9 +21,9 @@ public enum FieldTypeEnum {
     UINT32(Descriptors.FieldDescriptor.JavaType.INT, Integer.class.getSimpleName()),
     ENUM(Descriptors.FieldDescriptor.JavaType.ENUM, Enum.class.getSimpleName()),
     SFIXED32(Descriptors.FieldDescriptor.JavaType.INT, Integer.class.getSimpleName()),
-    SFIXED64(Descriptors.FieldDescriptor.JavaType.LONG, Long.class.getSimpleName()),
+    SFIXED64(Descriptors.FieldDescriptor.JavaType.LONG, Integer.class.getSimpleName()),
     SINT32(Descriptors.FieldDescriptor.JavaType.INT, Integer.class.getSimpleName()),
-    SINT64(Descriptors.FieldDescriptor.JavaType.LONG, Long.class.getSimpleName());
+    SINT64(Descriptors.FieldDescriptor.JavaType.LONG, Integer.class.getSimpleName());
 
     private Descriptors.FieldDescriptor.JavaType javaType;
     private String javaSimpleType;
